main/requirements.txt
# fastapi == 0.95.1
# uvicorn == 0.22.0
pydantic == 1.10.7
prometheus-fastapi-instrumentator == 6.0.0
aiokafka == 0.8.0
